requirements.txt
-e .
aws-cdk.aws-athena==1.105.0
aws-cdk.aws-dms==1.105.0
aws-cdk.aws-ec2==1.105.0
aws-cdk.aws-emr==1.105.0
aws-cdk.aws-glue==1.105.0
aws-cdk.aws-iam==1.105.0
aws-cdk.aws-kinesis==1.105.0
aws-cdk.aws-kinesisfirehose==1.105.0
aws-cdk.aws-rds==1.105.0
aws-cdk.aws-redshift==1.105.0
aws-cdk.aws-s3==1.105.0
black==21.5b1
boto3==1.17.76
fake-web-events==0.2.5
pandas==1.2.4
psycopg2-binary==2.8.6
pytest==6.2.4
requests==2.25.1
dbt==0.19.1